move custom entities related code
