+++ internal/dataplane/kongstate/kongstate.go
@@ -1,9 +1,7 @@
 package kongstate
 
 import (
-	"context"
 	"crypto/sha256"
-	"encoding/json"
 	"fmt"
 	"strconv"
 	"strings"
@@ -676,222 +674,6 @@ func maybeLogKongIngressDeprecationError(logger logr.Logger, services []*corev1.
 			)
 		}
 	}
-}
-
-// FillCustomEntities fills custom entities in KongState.
-func (ks *KongState) FillCustomEntities(
-	logger logr.Logger,
-	s store.Storer,
-	failuresCollector *failures.ResourceFailuresCollector,
-	schemaGetter SchemaGetter,
-	workspace string,
-) {
-	entities := s.ListKongCustomEntities()
-	if len(entities) == 0 {
-		return
-	}
-	logger = logger.WithName("fillCustomEntities")
-
-	if ks.CustomEntities == nil {
-		ks.CustomEntities = map[string]*KongCustomEntityCollection{}
-	}
-	// Fetch relations between plugins and services/routes/consumers and store the pointer to translated Kong entities.
-	// Used for fetching entity referred by a custom entity and fill the ID of referred entity.
-	pluginRels := ks.getPluginRelatedEntitiesRef(s, logger)
-
-	for _, entity := range entities {
-		// reject the custom entity if its type is in "known" entity types that are already processed.
-		if IsKnownEntityType(entity.Spec.EntityType) {
-			failuresCollector.PushResourceFailure(
-				fmt.Sprintf("cannot use known entity type %s in custom entity", entity.Spec.EntityType),
-				entity,
-			)
-			continue
-		}
-		// Fetch the entity schema.
-		schema, err := ks.fetchEntitySchema(schemaGetter, entity.Spec.EntityType)
-		if err != nil {
-			failuresCollector.PushResourceFailure(
-				fmt.Sprintf("failed to fetch entity schema for entity type %s: %v", entity.Spec.EntityType, err),
-				entity,
-			)
-			continue
-		}
-		// Unmarshal fields of the entity.
-		var parsedEntity map[string]any
-		if err = json.Unmarshal(entity.Spec.Fields.Raw, &parsedEntity); err != nil {
-			failuresCollector.PushResourceFailure(fmt.Sprintf("failed to unmarshal fields of entity: %v", err), entity)
-			continue
-		}
-		// Fill the "foreign" fields if the entity has such fields referencing services/routes/consumers.
-		// First Find out possible foreign field combinations attached to the KCE resource.
-		foreignFieldCombinations := ks.findCustomEntityForeignFields(logger, entity, schema, pluginRels, workspace)
-		// If there are any foreign fields, generate one entity per each foreign entity combination.
-		if len(foreignFieldCombinations) > 0 {
-			for _, combination := range foreignFieldCombinations {
-				// REVIEW: We do not have a built-in or common method to deep copy a `map[string]any`,
-				// so we decode it from the original JSON each time when generate an entity.
-				var parsedEntity map[string]any
-				if err = json.Unmarshal(entity.Spec.Fields.Raw, &parsedEntity); err != nil {
-					continue
-				}
-				generatedEntity := CustomEntity{
-					K8sKongCustomEntity: entity,
-					ForeignEntityIDs:    make(map[kong.EntityType]string),
-				}
-				// fill the fields referring to foreign entities.
-				for _, foreignField := range combination {
-					parsedEntity[foreignField.fieldName] = map[string]any{
-						"id": foreignField.foreignEntityID,
-					}
-					// save the referred foreign entity IDs for sorting.
-					generatedEntity.ForeignEntityIDs[foreignField.foreignEntityType] = foreignField.foreignEntityID
-				}
-				generatedEntity.Object = parsedEntity
-				ks.addCustomEntity(entity.Spec.EntityType, schema, generatedEntity)
-			}
-		} else {
-			ks.addCustomEntity(
-				entity.Spec.EntityType, schema,
-				CustomEntity{
-					Object:              parsedEntity,
-					K8sKongCustomEntity: entity,
-				},
-			)
-		}
-	}
-
-	ks.sortCustomEntities()
-}
-
-// addCustomEntity adds a custom entity into the collection of its type.
-func (ks *KongState) addCustomEntity(entityType string, schema EntitySchema, e CustomEntity) {
-	// Put the entity into the custom collection to store the entities of its type.
-	if _, ok := ks.CustomEntities[entityType]; !ok {
-		ks.CustomEntities[entityType] = &KongCustomEntityCollection{
-			Schema: schema,
-		}
-	}
-	collection := ks.CustomEntities[entityType]
-	collection.Entities = append(collection.Entities, e)
-}
-
-// fetchEntitySchema fetches schema of an entity by its type and stores the schema in its custom entity collection
-// as a cache to avoid excessive calling of Kong admin APIs.
-func (ks *KongState) fetchEntitySchema(schemaGetter SchemaGetter, entityType string) (EntitySchema, error) {
-	collection, ok := ks.CustomEntities[entityType]
-	if ok {
-		return collection.Schema, nil
-	}
-	// Use `context.Background()` here because `BuildKongConfig` does not provide a context.
-	schema, err := schemaGetter.Get(context.Background(), entityType)
-	if err != nil {
-		return EntitySchema{}, err
-	}
-	return ExtractEntityFieldDefinitions(schema), nil
-}
-
-func (ks *KongState) findCustomEntityRelatedPlugin(k8sEntity *kongv1alpha1.KongCustomEntity) (string, bool) {
-	// Find referred entity via the plugin in its spec.parentRef.
-	// Then we can fetch the referred service/route/consumer from the reference relations of the plugin.
-	parentRef := k8sEntity.Spec.ParentRef
-	var namespace string
-	// Abort if the parentRef is empty or does not refer to a plugin.
-	if parentRef == nil ||
-		(parentRef.Group == nil || *parentRef.Group != kongv1alpha1.GroupVersion.Group) {
-		return "", false
-	}
-	if parentRef.Kind == nil || (*parentRef.Kind != "KongPlugin" && *parentRef.Kind != "KongClusterPlugin") {
-		return "", false
-	}
-	// Extract the plugin key to get the plugin relations.
-	if parentRef.Namespace == nil || *parentRef.Namespace == "" {
-		namespace = k8sEntity.Namespace
-	} else {
-		namespace = *parentRef.Namespace
-	}
-	return namespace + ":" + parentRef.Name, true
-}
-
-func (ks *KongState) findCustomEntityForeignFields(
-	logger logr.Logger,
-	k8sEntity *kongv1alpha1.KongCustomEntity,
-	schema EntitySchema,
-	pluginRelEntities PluginRelatedEntitiesRefs,
-	workspace string,
-) [][]entityForeignFieldValue {
-	pluginKey, ok := ks.findCustomEntityRelatedPlugin(k8sEntity)
-	if !ok {
-		return nil
-	}
-	// Get the relations with other entities of the plugin.
-	rels, ok := pluginRelEntities.RelatedEntities[pluginKey]
-	if !ok {
-		return nil
-	}
-
-	var (
-		foreignRelations      util.ForeignRelations
-		foreignServiceFields  []string
-		foreignRouteFields    []string
-		foreignConsumerFields []string
-	)
-
-	ret := [][]entityForeignFieldValue{}
-	for fieldName, field := range schema.Fields {
-		if field.Type != EntityFieldTypeForeign {
-			continue
-		}
-		switch field.Reference {
-		case string(kong.EntityTypeServices):
-			foreignServiceFields = append(foreignServiceFields, fieldName)
-			foreignRelations.Service = getServiceIDFromPluginRels(logger, rels, pluginRelEntities.RouteAttachedService, workspace)
-		case string(kong.EntityTypeRoutes):
-			foreignRouteFields = append(foreignRouteFields, fieldName)
-			foreignRelations.Route = lo.FilterMap(rels.Routes, func(r *Route, _ int) (string, bool) {
-				if err := r.FillID(workspace); err != nil {
-					return "", false
-				}
-				return *r.ID, true
-			})
-		case string(kong.EntityTypeConsumers):
-			foreignConsumerFields = append(foreignConsumerFields, fieldName)
-			foreignRelations.Consumer = lo.FilterMap(rels.Consumers, func(c *Consumer, _ int) (string, bool) {
-				if err := c.FillID(workspace); err != nil {
-					return "", false
-				}
-				return *c.ID, true
-			})
-		} // end of switch
-	}
-
-	for _, combination := range foreignRelations.GetCombinations() {
-		foreignFieldValues := []entityForeignFieldValue{}
-		for _, fieldName := range foreignServiceFields {
-			foreignFieldValues = append(foreignFieldValues, entityForeignFieldValue{
-				fieldName:         fieldName,
-				foreignEntityType: kong.EntityTypeServices,
-				foreignEntityID:   combination.Service,
-			})
-		}
-		for _, fieldName := range foreignRouteFields {
-			foreignFieldValues = append(foreignFieldValues, entityForeignFieldValue{
-				fieldName:         fieldName,
-				foreignEntityType: kong.EntityTypeRoutes,
-				foreignEntityID:   combination.Route,
-			})
-		}
-		for _, fieldName := range foreignConsumerFields {
-			foreignFieldValues = append(foreignFieldValues, entityForeignFieldValue{
-				fieldName:         fieldName,
-				foreignEntityType: kong.EntityTypeConsumers,
-				foreignEntityID:   combination.Consumer,
-			})
-		}
-		ret = append(ret, foreignFieldValues)
-	}
-
-	return ret
 }
 
 // getServiceIDFromPluginRels returns the ID of the services which a plugin refers to in RelatedEntitiesRef.
