+++ internal/dataplane/kongstate/kongstate_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/go-logr/logr/testr"
 	"github.com/go-logr/zapr"
 	"github.com/kong/go-kong/kong"
-	"github.com/kong/go-kong/kong/custom"
 	"github.com/samber/lo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -1526,389 +1525,4 @@ func (s *fakeSchemaGetter) Get(_ context.Context, entityType string) (kong.Schem
 		return nil, fmt.Errorf("schema not found")
 	}
 	return schema, nil
-}
-
-func TestKongState_FillCustomEntities(t *testing.T) {
-	customEntityTypeMeta := metav1.TypeMeta{
-		APIVersion: kongv1alpha1.GroupVersion.Group + "/" + kongv1alpha1.GroupVersion.Version,
-		Kind:       "KongCustomEntity",
-	}
-	kongService1 := kong.Service{
-		Name: kong.String("service1"),
-		ID:   kong.String("service1"),
-	}
-	kongService2 := kong.Service{
-		Name: kong.String("service2"),
-		ID:   kong.String("service2"),
-	}
-
-	testCases := []struct {
-		name                        string
-		initialState                *KongState
-		customEntities              []*kongv1alpha1.KongCustomEntity
-		plugins                     []*kongv1.KongPlugin
-		schemas                     map[string]kong.Schema
-		expectedCustomEntities      map[string][]custom.Object
-		expectedTranslationFailures map[k8stypes.NamespacedName]string
-	}{
-		{
-			name:         "single custom entity",
-			initialState: &KongState{},
-			customEntities: []*kongv1alpha1.KongCustomEntity{
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "session-foo",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "sessions",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"name":"session1"}`),
-						},
-					},
-				},
-			},
-			schemas: map[string]kong.Schema{
-				"sessions": {
-					"fields": []interface{}{
-						map[string]interface{}{
-							"name": map[string]interface{}{
-								"type":     "string",
-								"required": true,
-							},
-						},
-					},
-				},
-			},
-			expectedCustomEntities: map[string][]custom.Object{
-				"sessions": {
-					{
-						"name": "session1",
-					},
-				},
-			},
-		},
-		{
-			name:         "custom entity with unknown type",
-			initialState: &KongState{},
-			customEntities: []*kongv1alpha1.KongCustomEntity{
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "session-foo",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "sessions",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"name":"session1"}`),
-						},
-					},
-				},
-			},
-			expectedTranslationFailures: map[k8stypes.NamespacedName]string{
-				{
-					Namespace: "default",
-					Name:      "session-foo",
-				}: "failed to fetch entity schema for entity type sessions: schema not found",
-			},
-		},
-		{
-			name:         "multiple custom entities with same type",
-			initialState: &KongState{},
-			customEntities: []*kongv1alpha1.KongCustomEntity{
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "session-foo",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "sessions",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"name":"session-foo"}`),
-						},
-					},
-				},
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "session-bar",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "sessions",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"name":"session-bar"}`),
-						},
-					},
-				},
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default-1",
-						Name:      "session-foo",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "sessions",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"name":"session-foo-1"}`),
-						},
-					},
-				},
-			},
-			schemas: map[string]kong.Schema{
-				"sessions": {
-					"fields": []interface{}{
-						map[string]interface{}{
-							"name": map[string]interface{}{
-								"type":     "string",
-								"required": true,
-							},
-						},
-					},
-				},
-			},
-			expectedCustomEntities: map[string][]custom.Object{
-				// Should be sorted by original KCE namespace/name.
-				"sessions": {
-					{
-						// from default/bar
-						"name": "session-bar",
-					},
-					{
-						// from default/foo
-						"name": "session-foo",
-					},
-					{
-						// from default-1/foo
-						"name": "session-foo-1",
-					},
-				},
-			},
-		},
-		{
-			name: "custom entities with reference to other entities (services)",
-			initialState: &KongState{
-				Services: []Service{
-					{
-						Service: kongService1,
-						K8sServices: map[string]*corev1.Service{
-							"default/service1": {
-								ObjectMeta: metav1.ObjectMeta{
-									Namespace: "default",
-									Name:      "service1",
-									Annotations: map[string]string{
-										annotations.AnnotationPrefix + annotations.PluginsKey: "degraphql-1",
-									},
-								},
-							},
-						},
-					}, // Service: service1
-				}, // Services
-			},
-			customEntities: []*kongv1alpha1.KongCustomEntity{
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "degraphql-1",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "degraphql_routes",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"uri":"/api/me"}`),
-						},
-						ParentRef: &kongv1alpha1.ObjectReference{
-							Group: kong.String(kongv1.GroupVersion.Group),
-							Kind:  kong.String("KongPlugin"),
-							Name:  "degraphql-1",
-						},
-					},
-				},
-			},
-			plugins: []*kongv1.KongPlugin{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "degraphql-1",
-					},
-					PluginName: "degraphql",
-				},
-			},
-			schemas: map[string]kong.Schema{
-				"degraphql_routes": {
-					"fields": []interface{}{
-						map[string]interface{}{
-							"uri": map[string]interface{}{
-								"type":     "string",
-								"required": true,
-							},
-						},
-						map[string]interface{}{
-							"service": map[string]interface{}{
-								"type":      "foreign",
-								"reference": "services",
-							},
-						},
-					},
-				},
-			},
-			expectedCustomEntities: map[string][]custom.Object{
-				"degraphql_routes": {
-					{
-						"uri": "/api/me",
-						"service": map[string]interface{}{
-							// ID generated from Kong service "service1" in workspace "".
-							"id": "service1",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "custom entity attached to multiple services via plugin",
-			initialState: &KongState{
-				Services: []Service{
-					{
-						Service: kongService1,
-						K8sServices: map[string]*corev1.Service{
-							"default/service1": {
-								ObjectMeta: metav1.ObjectMeta{
-									Namespace: "default",
-									Name:      "service1",
-									Annotations: map[string]string{
-										annotations.AnnotationPrefix + annotations.PluginsKey: "degraphql-1",
-									},
-								},
-							},
-						},
-					}, // Service: service1
-					{
-						Service: kongService2,
-						K8sServices: map[string]*corev1.Service{
-							"default/service2": {
-								ObjectMeta: metav1.ObjectMeta{
-									Namespace: "default",
-									Name:      "service2",
-									Annotations: map[string]string{
-										annotations.AnnotationPrefix + annotations.PluginsKey: "degraphql-1",
-									},
-								},
-							},
-						},
-					}, // Service: service2
-				}, // Services
-			},
-			customEntities: []*kongv1alpha1.KongCustomEntity{
-				{
-					TypeMeta: customEntityTypeMeta,
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "degraphql-1",
-					},
-					Spec: kongv1alpha1.KongCustomEntitySpec{
-						EntityType:     "degraphql_routes",
-						ControllerName: annotations.DefaultIngressClass,
-						Fields: apiextensionsv1.JSON{
-							Raw: []byte(`{"uri":"/api/me"}`),
-						},
-						ParentRef: &kongv1alpha1.ObjectReference{
-							Group: kong.String(kongv1.GroupVersion.Group),
-							Kind:  kong.String("KongPlugin"),
-							Name:  "degraphql-1",
-						},
-					},
-				},
-			},
-			plugins: []*kongv1.KongPlugin{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "default",
-						Name:      "degraphql-1",
-					},
-					PluginName: "degraphql",
-				},
-			},
-			schemas: map[string]kong.Schema{
-				"degraphql_routes": {
-					"fields": []interface{}{
-						map[string]interface{}{
-							"uri": map[string]interface{}{
-								"type":     "string",
-								"required": true,
-							},
-						},
-						map[string]interface{}{
-							"service": map[string]interface{}{
-								"type":      "foreign",
-								"reference": "services",
-							},
-						},
-					},
-				},
-			},
-			expectedCustomEntities: map[string][]custom.Object{
-				"degraphql_routes": {
-					{
-						"uri": "/api/me",
-						"service": map[string]interface{}{
-							// ID generated from Kong service "service1" in workspace "".
-							"id": "service1",
-						},
-					},
-					{
-						"uri": "/api/me",
-						"service": map[string]interface{}{
-							// ID generated from Kong service "service2" in workspace "".
-							"id": "service2",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			s, err := store.NewFakeStore(store.FakeObjects{
-				KongCustomEntities: tc.customEntities,
-				KongPlugins:        tc.plugins,
-			})
-			require.NoError(t, err)
-			failuresCollector := failures.NewResourceFailuresCollector(logr.Discard())
-
-			ks := tc.initialState
-			ks.FillCustomEntities(
-				logr.Discard(), s,
-				failuresCollector,
-				&fakeSchemaGetter{schemas: tc.schemas}, "",
-			)
-			for entityType, expectedObjectList := range tc.expectedCustomEntities {
-				require.NotNil(t, ks.CustomEntities[entityType])
-				objectList := lo.Map(ks.CustomEntities[entityType].Entities, func(e CustomEntity, _ int) custom.Object {
-					return e.Object
-				})
-				require.Equal(t, expectedObjectList, objectList)
-			}
-
-			translationFailures := failuresCollector.PopResourceFailures()
-			for nsName, message := range tc.expectedTranslationFailures {
-				hasError := lo.ContainsBy(translationFailures, func(f failures.ResourceFailure) bool {
-					fmt.Println(f.Message())
-					return f.Message() == message && lo.ContainsBy(f.CausingObjects(), func(o client.Object) bool {
-						return o.GetNamespace() == nsName.Namespace && o.GetName() == nsName.Name
-					})
-				})
-				require.Truef(t, hasError, "translation error for KongCustomEntity %s not found", nsName)
-			}
-		})
-	}
 }
